Finished the most basic functionality. It will properly show the amount for each day

+++ back-end/src/main/java/org/example/backend/savings_logic/SavingsService.java
@@ -2,9 +2,10 @@ package org.example.backend.savings_logic;
 
 import jakarta.persistence.criteria.CriteriaBuilder;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
 public interface SavingsService {
-    public List<Integer> computeAmountPerDayForUserIdAndBankAccountId(Integer userId, Integer bankAccountId);
+    public List<BigDecimal> computeAmountPerDayForUserIdAndBankAccountId(Integer userId, Integer bankAccountId);
 }
